+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/qc/DetailedQcItem.java
@@ -1,5 +1,7 @@
 package uk.ac.bbsrc.tgac.miso.core.data.qc;
 
+import com.eaglegenomics.simlims.core.User;
+
 import uk.ac.bbsrc.tgac.miso.core.data.DetailedQcStatus;
 
 public interface DetailedQcItem {
@@ -11,5 +13,9 @@ public interface DetailedQcItem {
   public String getDetailedQcStatusNote();
 
   public void setDetailedQcStatusNote(String detailedQcStatusNote);
+
+  public User getQcUser();
+
+  public void setQcUser(User qcUser);
 
 }